+++ temba/templates/tests.py
@@ -14,7 +14,6 @@ from temba.tests import CRUDLTestMixin, TembaTest
 
 from .models import Template, TemplateTranslation
 from .tasks import refresh_templates
-from .whatsapp import _extract_params, parse_language
 
 
 class TemplateTest(TembaTest):
@@ -27,7 +26,16 @@ class TemplateTest(TembaTest):
             external_id="1234",
             external_locale="en_US",
             namespace="",
-            components=[{"type": "body", "name": "body", "content": "Hello {{1}}", "params": [{"type": "text"}]}],
+            components=[
+                {
+                    "type": "body",
+                    "name": "body",
+                    "content": "Hello {{1}}",
+                    "variables": {"1": 0},
+                    "params": [{"type": "text"}],
+                }
+            ],
+            variables=[{"type": "text"}],
         )
         tt2 = TemplateTranslation.get_or_create(
             self.channel,
@@ -37,7 +45,16 @@ class TemplateTest(TembaTest):
             external_id="5678",
             external_locale="fr_FR",
             namespace="",
-            components=[{"type": "body", "name": "body", "content": "Bonjour {{1}}", "params": [{"type": "text"}]}],
+            components=[
+                {
+                    "type": "body",
+                    "name": "body",
+                    "content": "Bonjour {{1}}",
+                    "variables": {"1": 0},
+                    "params": [{"type": "text"}],
+                }
+            ],
+            variables=[{"type": "text"}],
         )
 
         self.assertEqual(tt1.template, tt2.template)
@@ -51,7 +68,16 @@ class TemplateTest(TembaTest):
             external_id="5678",
             external_locale="fr_FR",
             namespace="foo_namespace",
-            components=[{"type": "body", "name": "body", "content": "Salut {{1}}", "params": [{"type": "text"}]}],
+            components=[
+                {
+                    "type": "body",
+                    "name": "body",
+                    "content": "Salut {{1}}",
+                    "variables": {"1": 0},
+                    "params": [{"type": "text"}],
+                }
+            ],
+            variables=[{"type": "text"}],
         )
 
         self.assertTrue(tt3.template.modified_on > modified_on)
@@ -66,7 +92,7 @@ class TemplateTest(TembaTest):
         tt2.refresh_from_db()
         self.assertFalse(tt2.is_active)
 
-    @patch("temba.templates.models.Template.update_local")
+    @patch("temba.templates.models.TemplateTranslation.update_local")
     @patch("temba.channels.types.dialog360.Dialog360Type.fetch_templates")
     @patch("temba.channels.types.dialog360_legacy.Dialog360LegacyType.fetch_templates")
     def test_refresh_task(self, mock_d3_fetch_templates, mock_d3c_fetch_templates, mock_update_local):
@@ -318,7 +344,7 @@ class TemplateTest(TembaTest):
                     {
                         "type": "BUTTONS",
                         "buttons": [
-                            {"type": "QUICK_REPLY", "text": "Yes"},
+                            {"type": "QUICK_REPLY", "text": "Yes {{1}}"},
                             {"type": "QUICK_REPLY", "text": "No"},
                             {"type": "PHONE_NUMBER", "text": "Call center", "phone_number": "+1234"},
                             {
@@ -355,7 +381,7 @@ class TemplateTest(TembaTest):
             },
         ]
 
-        Template.update_local(channel, wa_templates)
+        TemplateTranslation.update_local(channel, wa_templates)
 
         self.assertEqual(8, Template.objects.filter(org=self.org).count())
         self.assertEqual(10, TemplateTranslation.objects.filter(channel=channel).count())
@@ -379,11 +405,13 @@ class TemplateTest(TembaTest):
                     "type": "body",
                     "name": "body",
                     "content": "Goodbye {{1}}, see you on {{2}}. See you later {{1}}",
+                    "variables": {"1": 0, "2": 1},
                     "params": [{"type": "text"}, {"type": "text"}],
                 }
             ],
             ct.components,
         )
+        self.assertEqual([{"type": "text"}, {"type": "text"}], ct.variables)
 
         ct = TemplateTranslation.objects.get(template__name="workout_activity", is_active=True)
         self.assertEqual("eng", ct.locale)
@@ -395,18 +423,27 @@ class TemplateTest(TembaTest):
                     "type": "header",
                     "name": "header",
                     "content": "Workout challenge week extra points!",
+                    "variables": {},
                     "params": [],
                 },
                 {
                     "type": "body",
                     "name": "body",
                     "content": "Hey {{1}}, Week {{2}} workout is out now. Get your discount of {{3}} for the next workout by sharing this program to 3 people.",
+                    "variables": {"1": 0, "2": 1, "3": 2},
                     "params": [{"type": "text"}, {"type": "text"}, {"type": "text"}],
                 },
-                {"type": "footer", "name": "footer", "content": "Remember to drink water.", "params": []},
+                {
+                    "type": "footer",
+                    "name": "footer",
+                    "content": "Remember to drink water.",
+                    "variables": {},
+                    "params": [],
+                },
             ],
             ct.components,
         )
+        self.assertEqual([{"type": "text"}, {"type": "text"}, {"type": "text"}], ct.variables)
 
         ct = TemplateTranslation.objects.get(template__name="invalid_component", is_active=True)
         self.assertEqual("fra", ct.locale)
@@ -420,8 +457,8 @@ class TemplateTest(TembaTest):
         self.assertEqual("foo_namespace", ct.namespace)
         self.assertEqual(
             [
-                {"type": "body", "name": "body", "content": "", "params": []},
-                {"type": "footer", "name": "footer", "content": "", "params": []},
+                {"type": "body", "name": "body", "content": "", "variables": {}, "params": []},
+                {"type": "footer", "name": "footer", "content": "", "variables": {}, "params": []},
             ],
             ct.components,
         )
@@ -432,21 +469,35 @@ class TemplateTest(TembaTest):
         self.assertEqual("foo_namespace", ct.namespace)
         self.assertEqual(
             [
-                {"type": "header", "name": "header", "content": "", "params": []},
+                {"type": "header", "name": "header", "content": "", "variables": {}, "params": []},
                 {
                     "type": "body",
                     "name": "body",
                     "content": "Sorry your order {{1}} took longer to deliver than expected.\nWe'll notify you about updates in the next {{2}} days.\n\nDo you have more question?",
+                    "variables": {"1": 0, "2": 1},
                     "params": [{"type": "text"}, {"type": "text"}],
                 },
-                {"type": "footer", "name": "footer", "content": "Thanks for your patience", "params": []},
-                {"type": "button/quick_reply", "name": "button.0", "content": "Yes", "params": []},
-                {"type": "button/quick_reply", "name": "button.1", "content": "No", "params": []},
+                {
+                    "type": "footer",
+                    "name": "footer",
+                    "content": "Thanks for your patience",
+                    "variables": {},
+                    "params": [],
+                },
+                {
+                    "type": "button/quick_reply",
+                    "name": "button.0",
+                    "content": "Yes {{1}}",
+                    "variables": {"1": 2},
+                    "params": [{"type": "text"}],
+                },
+                {"type": "button/quick_reply", "name": "button.1", "content": "No", "variables": {}, "params": []},
                 {
                     "type": "button/phone_number",
                     "name": "button.2",
                     "content": "+1234",
                     "display": "Call center",
+                    "variables": {},
                     "params": [],
                 },
                 {
@@ -454,6 +505,7 @@ class TemplateTest(TembaTest):
                     "name": "button.3",
                     "content": r"https:\/\/example.com\/?wa_customer={{1}}",
                     "display": "Check website",
+                    "variables": {"1": 3},
                     "params": [{"type": "text"}],
                 },
                 {
@@ -461,11 +513,13 @@ class TemplateTest(TembaTest):
                     "name": "button.4",
                     "content": r"https:\/\/example.com\/help",
                     "display": "Check website",
+                    "variables": {},
                     "params": [],
                 },
             ],
             ct.components,
         )
+        self.assertEqual([{"type": "text"}, {"type": "text"}, {"type": "text"}, {"type": "text"}], ct.variables)
 
     def test_update_local_templates_dialog360(self):
         # no namespace in channel config
@@ -642,7 +696,7 @@ class TemplateTest(TembaTest):
             },
         ]
 
-        Template.update_local(channel, D3_templates_data)
+        TemplateTranslation.update_local(channel, D3_templates_data)
 
         self.assertEqual(10, Template.objects.filter(org=self.org).count())
         self.assertEqual(12, TemplateTranslation.objects.filter(channel=channel).count())
@@ -696,7 +750,16 @@ class TemplateTranslationCRUDLTest(CRUDLTestMixin, TembaTest):
             external_id="1234",
             external_locale="en_US",
             namespace="foo_namespace",
-            components=[{"type": "body", "name": "body", "content": "Hello {{1}}", "params": [{"type": "text"}]}],
+            components=[
+                {
+                    "type": "body",
+                    "name": "body",
+                    "content": "Hello {{1}}",
+                    "variables": {"1": 0},
+                    "params": [{"type": "text"}],
+                }
+            ],
+            variables=[{"type": "text"}],
         )
         tt2 = TemplateTranslation.get_or_create(
             channel,
@@ -706,7 +769,16 @@ class TemplateTranslationCRUDLTest(CRUDLTestMixin, TembaTest):
             external_id="2345",
             external_locale="en_US",
             namespace="foo_namespace",
-            components=[{"type": "body", "name": "body", "content": "Goodbye {{1}}", "params": [{"type": "text"}]}],
+            components=[
+                {
+                    "type": "body",
+                    "name": "body",
+                    "content": "Goodbye {{1}}",
+                    "variables": {"1": 0},
+                    "params": [{"type": "text"}],
+                }
+            ],
+            variables=[{"type": "text"}],
         )
 
         # and one for another channel
@@ -718,7 +790,16 @@ class TemplateTranslationCRUDLTest(CRUDLTestMixin, TembaTest):
             external_id="5678",
             external_locale="en_US",
             namespace="foo_namespace",
-            components=[{"type": "body", "name": "body", "content": "Goodbye {{1}}", "params": [{"type": "text"}]}],
+            components=[
+                {
+                    "type": "body",
+                    "name": "body",
+                    "content": "Goodbye {{1}}",
+                    "variables": {"1": 0},
+                    "params": [{"type": "text"}],
+                }
+            ],
+            variables=[{"type": "text"}],
         )
 
         channel_url = reverse("templates.templatetranslation_channel", args=[channel.uuid])
@@ -731,15 +812,3 @@ class TemplateTranslationCRUDLTest(CRUDLTestMixin, TembaTest):
 
         response = self.client.get(reverse("templates.templatetranslation_channel", args=["1234567890-1234"]))
         self.assertEqual(404, response.status_code)
-
-
-class WhatsAppUtilsTest(TembaTest):
-    def test_parse_language(self):
-        self.assertEqual("eng", parse_language("en"))
-        self.assertEqual("eng-US", parse_language("en_US"))
-        self.assertEqual("fil", parse_language("fil"))
-
-    def test_extract_params(self):
-        self.assertEqual([{"type": "text"}, {"type": "text"}], _extract_params("Hi {{1}} how are you? {{2}}"))
-        self.assertEqual([{"type": "text"}, {"type": "text"}], _extract_params("Hi {{1}} how are you? {{2}} {{1}}"))
-        self.assertEqual([], _extract_params("Hi there."))